Removed `FAILED_APPLY_TRX` emitting
This was emitted on failed transaction but this actually is really a cancel block message meaning the full block should be discarded.

Having `FAILED_APPLY_TRX` was causing some actual issues because after it's called, we were cancelling the block leading to a panic within geth node. The panic was probably recovered correctly but there is no reason to make it happen.

So we now use only `CANCEL_BLOCK` signalling, also, `FirehoseContext.CancelBlock` now does not exit the block if no block is in progress.